refactor: implement granular sync indicators and improve data precision
- Move BigDecimalSerializer to a dedicated utility file.
- Add per-network sync tracking in Dashboard and Detail ViewModels.
- Create SyncPulseIndicator common component for real-time sync feedback.
- Update UI states to use BigDecimal for total balance calculations.
- Integrate sync indicators into AssetCard and BalanceRow components.

+++ app/src/main/java/com/example/nexuswallet/feature/wallet/data/local/entity/BitcoinBalanceEntity.kt
@@ -4,6 +4,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 import java.util.UUID
 
 @Entity(
@@ -24,6 +25,6 @@ data class BitcoinBalanceEntity(
     val address: String,
     val satoshis: String,
     val btc: String,
-    val usdValue: Double,
+    val usdValue: BigDecimal,
     val updatedAt: Long = System.currentTimeMillis()
 )
